fix(serial): correct two defects found reviewing the review fixes
Three reviewers over my own fix commit found two P1s I had introduced.

The firewall rule named the wrong address. r.SerialHostAddr was set from the
bind address, but the container dials the advertised host, and those differ on
exactly one runtime: Docker Desktop binds 127.0.0.1 and the container reaches
it through host.docker.internal. The emitted rule matched only the container's
own loopback while the real packet went to the Desktop gateway and hit the
DROP, so a strict run's device was silently unreachable there — the same
failure the destination-scoping was meant to prevent, reintroduced by scoping
to the wrong end. The manager-level stub discarded that argument entirely,
which is why no test could see it; it is forwarded now, and the distinction
between bind and advertised host is pinned where the two actually diverge.

The termios test did not compile for darwin. Termios.Cflag is uint32 on Linux
and uint64 on Darwin; the original code had explicit conversions, unconvert
flagged them as redundant against the Linux width, and removing them broke the
other platform. go vet on Linux cannot see this, so darwin cross-compilation is
now part of the check I run.

Also tightens the handshake budget the previous commit added. Clearing it on
the first byte was a weaker bar than intended: a single stray IAC disabled it
permanently, leaving the same slow-loris with one byte of extra effort. It is
now an absolute budget cleared only by a completed negotiation, a completed
com-port command, or payload reaching the device — a rolling per-byte deadline
would not have helped, since a peer can dribble one byte per window forever.
The timeout is injectable so the three cases are actually testable.

Writing those tests surfaced two of my own weak assertions. "Read returned an
error" is satisfied by the test's own read deadline, so it could not tell an
evicted peer from one that stopped waiting; it now requires a non-timeout
error. And a client-side Write succeeds into the local buffer after the peer
closes, so proving a session survived needs the device side observed, not the
writer's return value. Both were caught by mutation checks that initially
passed against deliberately broken implementations.

Unchanged and verified correct by review: the merge conflict resolution (all
tests from both sides present and passing), the termios extraction itself
(byte-identical to both originals), the accept-loop backoff, the rc.Close
ordering, and the approvedFromSpec extraction.

+++ internal/serialport/port_unix_test.go
@@ -321,9 +321,12 @@ func TestApplyFramingSetsTermiosBits(t *testing.T) {
 	}
 
 	t.Run("data bits select the CSIZE value", func(t *testing.T) {
-		for bits, want := range map[uint8]uint32{5: unix.CS5, 6: unix.CS6, 7: unix.CS7, 8: unix.CS8} {
+		// uint64 on both sides: Termios.Cflag is uint32 on Linux and uint64 on
+		// Darwin, so a width-specific comparison compiles on one and not the
+		// other — and `go vet` on Linux never sees the Darwin failure.
+		for bits, want := range map[uint8]uint64{5: unix.CS5, 6: unix.CS6, 7: unix.CS7, 8: unix.CS8} {
 			tio := bit(Settings{DataBits: bits})
-			if got := tio.Cflag & unix.CSIZE; got != want {
+			if got := uint64(tio.Cflag & unix.CSIZE); got != want {
 				t.Errorf("DataBits %d -> CSIZE %#x, want %#x", bits, got, want)
 			}
 		}
